python/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aristech-tts-client"
version = "1.0.0"
description = "A Python client library for the Aristech Text-to-Speech API"
authors = [{ name = "Martin Mende", email = "speech@aristech.de" }]
dependencies = ["grpcio", "protobuf"]
readme = "README.md"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
license = { text = "MIT License" }

[project.optional-dependencies]
dev = ["grpcio-tools", "mypy-protobuf", "python-dotenv"]

[project.urls]
Homepage = "https://github.com/aristech-de/tts-clients/tree/main/python"
Issues = "https://github.com/aristech-de/tts-clients/issues"
